+++ src/ploneintranet/userprofile/sync.py
@@ -10,6 +10,7 @@ from ploneintranet import api as pi_api
 from ploneintranet.core import ploneintranetCoreMessageFactory as _
 from Products.Five import BrowserView
 from Products.PluggableAuthService.interfaces.plugins import IPropertiesPlugin
+from Products.CMFPlone.utils import safe_unicode
 from transaction import commit
 from zope.component import adapter
 from zope.interface import alsoProvides
@@ -123,8 +124,8 @@ class UserPropertyManager(object):
                     data=value,
                     filename=u'portrait-%s.jpg' % self.context.username)
 
-            if value is not NO_VALUE and value != current_value:
-                setattr(self.context, property_name, value)
+            if value is not NO_VALUE and safe_unicode(value) != current_value:
+                setattr(self.context, property_name, safe_unicode(value))
                 changed = True
 
         if changed:
@@ -202,7 +203,7 @@ def create_membrane_profile(member):
         logger.info('Auto-creating membrane profile for {0}'.format(
             userid,
         ))
-        with api.env.adopt_roles(roles=('Manager', )):
+        with api.env.adopt_roles(roles=('Manager',)):
             profile = pi_api.userprofile.create(userid, approve=True)
             profile_manager = IUserProfileManager(profile)
             profile_manager.sync()
@@ -320,7 +321,8 @@ class WorkGroupPropertyManager(object):
         group = groups[0]
 
         def safe_set(data, key):
-            if data.get(key) and data[key] != getattr(self.context, key):
+            value = safe_unicode(data.get(key, ''))
+            if data.get(key) and value != getattr(self.context, key):
                 setattr(self.context, key, data[key])
                 return True
             return False
@@ -330,7 +332,11 @@ class WorkGroupPropertyManager(object):
 
         members = ext_source.getGroupMembers(self.context.canonical)
         if set(members) != set(self.context.members):
-            self.context.members = members
+            # Make sure all members are unicode
+            # AD allows unicode uids, we don't, and we break otherwise if such
+            # are included.
+            umembers = set([safe_unicode(x) for x in members])
+            self.context.members = umembers
             changed = True
 
         if changed: